ex pronto

+++ 23-08-repetição/23-08-ex2.c
@@ -4,7 +4,7 @@
 int main(void)
 {
     int P, N, I;
-    float R;
+    double R;
 
     printf("Coloque o Numero: ");
     scanf("%i",&N);
